Jenkins/jenkins-scenarios/scenario_02_test_master/requirements.txt
Flask==2.3.3
pytest==7.4.2
pytest-cov==4.1.0
testcontainers==3.7.1
psycopg2-binary==2.9.7
PyMySQL==1.1.0
redis==4.6.0
